Fixed the `Bosh::AzureCloud::Cloud.initialize` method's signature so that the 2nd argument is optional.
Note: Based on the changes made in commit #6c6ec5909b3635d5ea14a5f4dfa2263f2edf1468 (which added the 2nd arg), it seems that the ag was intended to be optional from the start. However, according to the method's signature, it was technically a required arg (despite that the method handles nil/missing values internally). Also, the following code already doesn't pass the 2nd arg:
- https://github.com/cloudfoundry/bosh-azure-cpi-release/blob/9754cf84f18be7ce0250518bd5a660a459470b56/src/bosh_azure_cpi/bin/bosh_azure_console#L29

Error Message:
> Missing argument 'api_version'. Required 2, 'options, api_version'